+++ packages/expo-updates/android/src/main/java/expo/modules/updates/loader/FileDownloader.java
@@ -23,6 +23,8 @@ import expo.modules.updates.db.entity.AssetEntity;
 import expo.modules.updates.launcher.NoDatabaseLauncher;
 import expo.modules.updates.manifest.Manifest;
 import expo.modules.updates.manifest.ManifestFactory;
+import expo.modules.updates.manifest.ManifestResponse;
+import okhttp3.Cache;
 import okhttp3.CacheControl;
 import okhttp3.Call;
 import okhttp3.Callback;
@@ -34,7 +36,7 @@ public class FileDownloader {
 
   private static final String TAG = FileDownloader.class.getSimpleName();
 
-  private static OkHttpClient sClient = new OkHttpClient.Builder().build();
+  private static OkHttpClient sClient;
 
   public interface FileDownloadCallback {
     void onFailure(Exception e);
@@ -51,8 +53,21 @@ public class FileDownloader {
     void onSuccess(AssetEntity assetEntity, boolean isNew);
   }
 
-  public static void downloadFileToPath(Request request, final File destination, final FileDownloadCallback callback) {
-    downloadData(request, new Callback() {
+  private static OkHttpClient getClient(Context context) {
+    if (sClient == null) {
+      sClient = new OkHttpClient.Builder().cache(getCache(context)).build();
+    }
+    return sClient;
+  }
+
+  private static Cache getCache(Context context) {
+    int cacheSize = 50 * 1024 * 1024; // 50 MiB
+    final File directory = new File(context.getCacheDir(), "okhttp");
+    return new Cache(directory, cacheSize);
+  }
+
+  public static void downloadFileToPath(Request request, final File destination, final Context context, final FileDownloadCallback callback) {
+    downloadData(request, context, new Callback() {
       @Override
       public void onFailure(Call call, IOException e) {
         callback.onFailure(e);
@@ -80,7 +95,7 @@ public class FileDownloader {
 
   public static void downloadManifest(final UpdatesConfiguration configuration, final Context context, final ManifestDownloadCallback callback) {
     try {
-      downloadData(setHeadersForManifestUrl(configuration, context), new Callback() {
+      downloadData(setHeadersForManifestUrl(configuration, context), context, new Callback() {
         @Override
         public void onFailure(Call call, IOException e) {
           callback.onFailure("Failed to download manifest from URL: " + configuration.getUpdateUrl(), e);
@@ -119,6 +134,7 @@ public class FileDownloader {
               Crypto.verifyPublicRSASignature(
                   manifestString,
                   signature,
+                  context,
                   new Crypto.RSASignatureListener() {
                     @Override
                     public void onError(Exception e, boolean isNetworkError) {
@@ -130,7 +146,7 @@ public class FileDownloader {
                       if (isValid) {
                         try {
                           preManifest.put("isVerified", true);
-                          Manifest manifest = ManifestFactory.getManifest(preManifest, configuration);
+                          Manifest manifest = ManifestFactory.getManifest(preManifest, new ManifestResponse(response), configuration);
                           callback.onSuccess(manifest);
                         } catch (JSONException e) {
                           callback.onFailure("Failed to parse manifest data", e);
@@ -143,7 +159,7 @@ public class FileDownloader {
               );
             } else {
               preManifest.put("isVerified", false);
-              Manifest manifest = ManifestFactory.getManifest(preManifest, configuration);
+              Manifest manifest = ManifestFactory.getManifest(preManifest, new ManifestResponse(response), configuration);
               callback.onSuccess(manifest);
             }
           } catch (Exception e) {
@@ -156,7 +172,7 @@ public class FileDownloader {
     }
   }
 
-  public static void downloadAsset(final AssetEntity asset, File destinationDirectory, UpdatesConfiguration configuration, final AssetDownloadCallback callback) {
+  public static void downloadAsset(final AssetEntity asset, File destinationDirectory, UpdatesConfiguration configuration, Context context, final AssetDownloadCallback callback) {
     if (asset.url == null) {
       callback.onFailure(new Exception("Could not download asset " + asset.key + " with no URL"), asset);
       return;
@@ -170,7 +186,7 @@ public class FileDownloader {
       callback.onSuccess(asset, false);
     } else {
       try {
-        downloadFileToPath(setHeadersForUrl(asset.url, configuration), path, new FileDownloadCallback() {
+        downloadFileToPath(setHeadersForUrl(asset.url, configuration), path, context, new FileDownloadCallback() {
           @Override
           public void onFailure(Exception e) {
             callback.onFailure(e, asset);
@@ -190,18 +206,18 @@ public class FileDownloader {
     }
   }
 
-  public static void downloadData(Request request, Callback callback) {
-    downloadData(request, callback, false);
+  public static void downloadData(Request request, Context context, Callback callback) {
+    downloadData(request, callback, context, false);
   }
 
-  private static void downloadData(final Request request, final Callback callback, final boolean isRetry) {
-    sClient.newCall(request).enqueue(new Callback() {
+  private static void downloadData(final Request request, final Callback callback, Context context, final boolean isRetry) {
+    getClient(context).newCall(request).enqueue(new Callback() {
       @Override
       public void onFailure(Call call, IOException e) {
         if (isRetry) {
           callback.onFailure(call, e);
         } else {
-          downloadData(request, callback, true);
+          downloadData(request, callback, context, true);
         }
       }
 
@@ -251,7 +267,7 @@ public class FileDownloader {
     return requestBuilder.build();
   }
 
-  private static Request setHeadersForManifestUrl(UpdatesConfiguration configuration, Context context) {
+  /* package */ static Request setHeadersForManifestUrl(UpdatesConfiguration configuration, Context context) {
     Request.Builder requestBuilder = new Request.Builder()
             .url(configuration.getUpdateUrl().toString())
             .header("Accept", "application/expo+json,application/json")
@@ -260,8 +276,12 @@ public class FileDownloader {
             .header("Expo-Updates-Environment", "BARE")
             .header("Expo-JSON-Error", "true")
             // as of 2020-11-25, the EAS Update alpha returns an error if Expo-Accept-Signature: true is included in the request
-            .header("Expo-Accept-Signature", String.valueOf(configuration.usesLegacyManifest()))
-            .cacheControl(CacheControl.FORCE_NETWORK);
+            .header("Expo-Accept-Signature", String.valueOf(configuration.usesLegacyManifest()));
+
+    // legacy manifest loads should ignore cache-control headers from the server and always load remotely
+    if (configuration.usesLegacyManifest()) {
+      requestBuilder = requestBuilder.cacheControl(CacheControl.FORCE_NETWORK);
+    }
 
     String runtimeVersion = configuration.getRuntimeVersion();
     String sdkVersion = configuration.getSdkVersion();