+++ packages/expo-updates/android/src/main/java/expo/modules/updates/manifest/NewManifest.java
@@ -3,6 +3,13 @@ package expo.modules.updates.manifest;
 import android.net.Uri;
 import android.util.Log;
 
+import androidx.annotation.Nullable;
+import expo.modules.structuredheaders.BooleanItem;
+import expo.modules.structuredheaders.Dictionary;
+import expo.modules.structuredheaders.ListElement;
+import expo.modules.structuredheaders.NumberItem;
+import expo.modules.structuredheaders.Parser;
+import expo.modules.structuredheaders.StringItem;
 import expo.modules.updates.UpdatesConfiguration;
 import expo.modules.updates.UpdatesUtils;
 import expo.modules.updates.db.entity.AssetEntity;
@@ -15,6 +22,7 @@ import org.json.JSONObject;
 import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.Map;
 import java.util.UUID;
 
 import static expo.modules.updates.loader.EmbeddedLoader.BUNDLE_FILENAME;
@@ -31,6 +39,8 @@ public class NewManifest implements Manifest {
   private JSONArray mAssets;
 
   private JSONObject mManifestJson;
+  private String mServerDefinedHeaders;
+  private String mManifestFilters;
 
   private NewManifest(JSONObject manifestJson,
                       UUID id,
@@ -38,7 +48,9 @@ public class NewManifest implements Manifest {
                       Date commitTime,
                       String runtimeVersion,
                       JSONObject launchAsset,
-                      JSONArray assets) {
+                      JSONArray assets,
+                      String serverDefinedHeaders,
+                      String manifestFilters) {
     mManifestJson = manifestJson;
     mId = id;
     mScopeKey = scopeKey;
@@ -46,9 +58,11 @@ public class NewManifest implements Manifest {
     mRuntimeVersion = runtimeVersion;
     mLaunchAsset = launchAsset;
     mAssets = assets;
+    mServerDefinedHeaders = serverDefinedHeaders;
+    mManifestFilters = manifestFilters;
   }
 
-  public static NewManifest fromManifestJson(JSONObject rootManifestJson, UpdatesConfiguration configuration) throws JSONException {
+  public static NewManifest fromManifestJson(JSONObject rootManifestJson, ManifestResponse httpResponse, UpdatesConfiguration configuration) throws JSONException {
     JSONObject manifestJson = rootManifestJson;
     if (manifestJson.has("manifest")) {
       manifestJson = manifestJson.getJSONObject("manifest");
@@ -67,7 +81,24 @@ public class NewManifest implements Manifest {
       commitTime = new Date();
     }
 
-    return new NewManifest(manifestJson, id, configuration.getScopeKey(), commitTime, runtimeVersion, launchAsset, assets);
+    String serverDefinedHeaders = httpResponse != null ? httpResponse.header("expo-server-defined-headers") : null;
+    String manifestFilters = httpResponse != null ? httpResponse.header("expo-manifest-filters") : null;
+
+    return new NewManifest(manifestJson, id, configuration.getScopeKey(), commitTime, runtimeVersion, launchAsset, assets, serverDefinedHeaders, manifestFilters);
+  }
+
+  public @Nullable JSONObject getServerDefinedHeaders() {
+    if (mServerDefinedHeaders == null) {
+      return null;
+    }
+    return headerDictionaryToJSONObject(mServerDefinedHeaders);
+  }
+
+  public @Nullable JSONObject getManifestFilters() {
+    if (mManifestFilters == null) {
+      return null;
+    }
+    return headerDictionaryToJSONObject(mManifestFilters);
   }
 
   public JSONObject getRawManifestJson() {
@@ -116,5 +147,25 @@ public class NewManifest implements Manifest {
 
   public boolean isDevelopmentMode() {
     return false;
+  }
+
+  /* package */ static @Nullable JSONObject headerDictionaryToJSONObject(String headerDictionary) {
+    JSONObject jsonObject = new JSONObject();
+    Parser parser = new Parser(headerDictionary);
+    try {
+      Dictionary filtersDictionary = parser.parseDictionary();
+      Map<String, ListElement<? extends Object>> map = filtersDictionary.get();
+      for (String key : map.keySet()) {
+        ListElement<? extends Object> element = map.get(key);
+        // ignore any dictionary entries whose type is not string, number, or boolean
+        if (element instanceof StringItem || element instanceof BooleanItem || element instanceof NumberItem) {
+          jsonObject.put(key, element.get());
+        }
+      }
+    } catch (expo.modules.structuredheaders.ParseException | JSONException e) {
+      Log.e(TAG, "Failed to parse manifest header content", e);
+      return null;
+    }
+    return jsonObject;
   }
 }