+++ packages/unimodules-permissions-interface/ios/UMPermissionsInterface/UMPermissionsMethodsDelegate.h
@@ -19,4 +19,3 @@
     withPermissionsManager:(id<UMPermissionsInterface>)permissionsManager;
 
 @end
-